+++ cyberdrop_dl/models/types.py
@@ -16,7 +16,15 @@ from pydantic import (
 
 from cyberdrop_dl.url_objects import AbsoluteHttpURL
 
-from .validators import bytesize_to_str, change_path_suffix, falsy_as_none, falsy_as_tuple, to_timedelta, to_yarl_url
+from .validators import (
+    bytesize_to_str,
+    change_path_suffix,
+    falsy_as_none,
+    falsy_as_tuple,
+    remove_duplicates,
+    to_timedelta,
+    to_yarl_url,
+)
 
 type LogLevel = Annotated[
     Literal["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"], StringConstraints(to_upper=True, strip_whitespace=True)
@@ -35,6 +43,7 @@ type Timedelta = Annotated[
         str, return_type=str, when_used="json"
     ),  # Serialize as str to save it as sexageximal (hh:mm:ss) instead of pydantic's ISO duration (PT1H5M26S)
 ]
+type RemoveDuplicates[T: tuple[str, ...]] = Annotated[T, AfterValidator(remove_duplicates)]
 
 
 class _HttpURL(AnyUrl):